Fixed tab prototypes and ft_strrchr

+++ include/tab.h
@@ -6,14 +6,14 @@
 /*   By: chlimous <chlimous@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/01/03 01:43:52 by chlimous          #+#    #+#             */
-/*   Updated: 2024/01/04 01:13:38 by chlimous         ###   ########.fr       */
+/*   Updated: 2024/01/07 21:46:42 by chlimous         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #ifndef TAB_H
 # define TAB_H
 
-void	ft_rev_int_tab(int *tab, unsigned int size);
-void	ft_rev_long_tab(long *tab, unsigned int size);
+void	ft_rev_int_tab(int *tab, size_t size);
+void	ft_rev_long_tab(long *tab, size_t size);
 
 #endif
